use the new implementation and adjustments
