+++ src/Controls/src/Core/Brush/Brush.cs
@@ -13,7 +13,7 @@ namespace Microsoft.Maui.Controls
 	[System.ComponentModel.TypeConverter(typeof(BrushTypeConverter))]
 	public abstract partial class Brush : Element
 	{
-		static readonly ICache<Color, ImmutableBrush> _cache = new CacheWithSwitch(51);
+		static readonly ICache<Color, ImmutableBrush> _cache = new Lru64ColorVectorInlineBrushCache(64);
 
 		public static implicit operator Brush(Paint paint)
 		{
